added backport compatibility marked previous event deprecated and added todo for removal in 1.19 moved new Events from AdvancementEvent to AdvancementEvents fixed code style
